Rejig simulator tests
