+++ x/programs/simulator/src/lib.rs
@@ -38,8 +38,5 @@ mod bindings {
 mod simulator;
 mod state;
 
-#[cfg(test)]
-mod tests;
-
 pub use simulator::Simulator;
 pub use state::SimpleState;
